requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bidict==0.23.1
boto3==1.37.25
botocore==1.37.25
certifi==2024.12.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.3
daphne==4.2.0
dataclasses-json==0.6.7
deepgram-sdk==4.1.0
deprecation==2.1.0
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==35.0.0
frozenlist==1.6.0
future==1.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
imageio-ffmpeg==0.6.0
incremental==24.7.2
jiter==0.10.0
jmespath==1.0.1
marshmallow==3.26.1
msgpack==1.1.0
multidict==6.4.4
mypy_extensions==1.1.0
mysqlclient==2.2.6
openai==1.84.0
packaging==24.2
pillow==11.1.0
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.2.0
requests==2.32.3
resend==2.5.1
s3transfer==0.11.4
service-identity==24.2.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
Twisted==25.5.0
txaio==23.1.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
zope.interface==7.2
